refactor: structure of tab
