+++ src/app/components/LeftMenu/Tab/SearchButton.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components/macro';
+
+export const SearchButton = styled.button`
+    background: transparent;
+    border: none;
+    background-image: url('./icon_search.svg');
+    background-position: center;
+    background-size: 40%;
+    background-repeat: no-repeat;
+    width: 20%;
+    height: 54px;
+`;
