+++ src/app/components/LeftMenu/Tab/SearchInput.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components/macro';
+
+export const SearchInput = styled.input`
+    width: 80%;
+    border: none;
+    padding: 15px 0 15px 15px;
+    border-radius: 0 0 15px 15px;
+
+    &::placeholder {
+        color: #8a8a8a;
+    }
+`;
