+++ src/app/components/LeftMenu/index.tsx
@@ -1,10 +1,6 @@
 import * as React from 'react';
 import styled from 'styled-components/macro';
 
-import { useDispatch } from 'react-redux';
-
-import { useHistory } from 'react-router-dom';
-
 import Tab from './Tab';
 
 const Box = styled.div`
@@ -22,9 +18,6 @@ const Box = styled.div`
 `;
 
 export default function LeftMenu() {
-
-    const dispatch = useDispatch();
-    const history = useHistory();
 
     return (
         <>
